FUCK YES
rictic you bastard, PriorityQueue.poll() removes the first element, and
it removed the best one before we even started searching.

also, made a mathematically accurate heuristic, sorta.  Uses the secant
method to solve for the number of timesteps taken to reach 7 tiles to
the right under a full run.  this turns out to be a slight overestimate
for some unknown reason.