Fixed early A* termination bugs
Similar to changes on the no-lock branch.  Heuristic cost is now allowed
to go negative, and we compare potential solutions by the heuristic cost
of their farthest-explored branch, not just the local A* cost of the
node, which can return an artificially low cost because the heuristic
cost is deliberately low.

I'm tired and have a splitting headache so I don't know whether the
above makes any sense to anybody.