This is much better-tuned.
We propagate death back to penalize the frames leading up to it; this is
expensive when we're about to fall in a hole, but it keeps us out.

Also, the heuristic cost now accounts for a column of stuff in front of
it.  I swear I tried this before, but I must have been doing something
stupid.  This simple rule seems to work.